requirements.txt
pyserial==3.4
gTTS==2.0.4